Implement cgmanager support
The cgmanager support is not optional;  If we wanted to we could make it
optional and simply not support the StartUnit and StartTransientUnit if
libcgmanager is not found.

Upon login, we are placed in a new per-session cgroup (i.e.
/user/$uid.user/c-$session.session/) which is delegated to us.

Actual configuration of cgroup rlimits (memory, cpusets, etc) is not yet
implemented.

Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>