Implement Manager:StopUnit and Scope:Abandon
First, stop setting removeonempty at login.  This is in particular
causing problems for upstart.  Upstart uses the same cgroup name
for several possible job states.  Using removeonempty means that
we have an unpredictable time at which the cgroup is removed.  So
when pre-start is over, post-stop may kick in, create the cgroup
(which exists), now autoremove kicks in, then post-stop fails to
move its child task into the cgroup.

StopUnit is called when /etc/systemd/logind.conf has
	KillUserProcesses=yes
It receives the scope (unique system-wide) to be stopped, and
kills that scope.  That means killing any tasks in the cgroup
or its subdirectories, and removing the directories themselves.
We first set removeonempty, then recursively kill subdirectories,
then killall tasks, then try to remove the directory.  Some kernels
have a race which can cause a child cgroup to not yet be listed as
removed by the time we get to removing the parent cgroup.  That
should not matter for us since we've set removeonempty, and the
cgroup will get removed eventually anyway.  However, in the case
where the cgroup was already empty, removeonempty would do nothing
until the next time the cgroup *becomes* empty, so we still need
to try to remove the directory by hand.  One of the other should
always succeed.

Abandon is sent differently than StopUnit - it is sent over a
session-specific path, to interface org.freedesktop.systemd1.Scope.
So shim needs to register the session-specific object path when it
starts.  This means that shim cannot exit until the session has
stopped.

If there is a way to tell the dbus system bus to start us back
up when that path is accessed (which does not persist across reboot)
then we could do that.  Or, we can tweak the exit-on-inactivity
code to extend timeout so long as there are open sessions.  For now
I simply commented it out.

Something should definately be done, as the shim code has been written
until now assuming it would not persist.  For instance the cgmanager
socket is never closed.  If there isn't a way to avoid keeping the
shim long-running, then we will need to address these matters.

Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>