Start on Manager:StopUnit and Scope:Abandon
Stop setting removeonempty at login.

StopUnit is called when /etc/systemd/logind.conf has
	KillUserProcesses=yes
It receives the scope (unique system-wide) to be stopped, and
kills that scope.  That means killing any tasks in the cgroup
or its subdirectories, and removing the directories themselves.
We first set removeonempty, then recursively kill subdirectories,
then killall tasks, then try to remove the directory.  Some kernels
have a race which can cause a child cgroup to not yet be listed as
removed by the time we get to removing the parent cgroup.  That
should not matter for us since we've set removeonempty, and the
cgroup will get removed eventually anyway.  However, in the case
where the cgroup was already empty, removeonempty would do nothing
until the next time the cgroup *becomes* empty, so we still need
to try to remove the directory by hand.  One of the other should
always succeed.

Abandon does not yet work.

Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>